[python] Complete re-organization and make eos-plot-{1d,2d} work again
