+++ src/scripts/eos-plot-1d
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-# Copyright (c) 2016 Danny van Dyk
+# Copyright (c) 2016, 2018 Danny van Dyk
 #
 # This file is part of the EOS project. EOS is free software;
 # you can redistribute it and/or modify it under the terms of the GNU General
@@ -15,7 +15,9 @@
 # this program; if not, write to the Free Software Foundation, Inc., 59 Temple
 # Place, Suite 330, Boston, MA  02111-1307  USA
 
-from eosdata import *
+import argparse
+import eos
+import os
 
 def main():
     parser = argparse.ArgumentParser(description='Plot 1D marginal distributions')
@@ -36,18 +38,18 @@ def main():
     # open HDF5 data file
     basename = os.path.basename(args.input)
     if basename.startswith('mcmc_'):
-        datafile = MCMCDataFile(args.input)
+        datafile = eos.data.MCMCDataFile(args.input)
     elif basename.startswith('pmc_monolithic'):
-        datafile = PMCDataFile(args.input)
+        datafile = eos.data.PMCDataFile(args.input)
     elif basename.startswith('unc'):
-        datafile = UncertaintyDataFile(args.input)
+        datafile = eos.data.UncertaintyDataFile(args.input)
 
     # forward options
     argvars = vars(args)
     options = { key: argvars[key] for key in ['xmin', 'xmax', 'kde', 'kde_bandwidth'] }
 
     # plot data
-    plotter = Plotter1D(datafile, args.output)
+    plotter = eos.plot.Plotter1D(datafile, args.output)
     plotter.histogram(args.index, **options)
 
     exit(0);
