+++ src/scripts/eos-plot-2d
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-# Copyright (c) 2016 Danny van Dyk
+# Copyright (c) 2016, 2018 Danny van Dyk
 #
 # This file is part of the EOS project. EOS is free software;
 # you can redistribute it and/or modify it under the terms of the GNU General
@@ -15,7 +15,9 @@
 # this program; if not, write to the Free Software Foundation, Inc., 59 Temple
 # Place, Suite 330, Boston, MA  02111-1307  USA
 
-from eosdata import *
+import argparse
+import eos
+import os
 
 def main():
     parser = argparse.ArgumentParser(description='Plot 2D marginal distributions')
@@ -37,16 +39,18 @@ def main():
     # open HDF5 data file
     basename = os.path.basename(args.input)
     if basename.startswith('mcmc_'):
-        datafile = MCMCDataFile(args.input)
+        datafile = eos.data.MCMCDataFile(args.input)
     elif basename.startswith('pmc_monolithic'):
-        datafile = PMCDataFile(args.input)
+        datafile = eos.data.PMCDataFile(args.input)
+    elif basename.startswith('unc'):
+        datafile = eos.data.UncertaintyDataFile(args.input)
 
     # forward options
     argvars = vars(args)
     options = { key: argvars[key] for key in ['xmin', 'xmax', 'ymin', 'ymax'] }
 
     # plot data
-    plotter = Plotter2D(datafile, args.output)
+    plotter = eos.plot.Plotter2D(datafile, args.output)
     plotter.histogram(args.xindex, args.yindex, **options)
 
     exit(0);
